add) 책 기록 화면 추가 -(찜, 내 책 id Long타입 변경)
1) RecordScreen.kt
내 책 목록 화면 추가.

2) RecordDetailScreen.kt
내 책에서 상세보기 추가. (표시 정보가 기존 Detail과 달라서 추가.)

3) RecordViewModel.kt, RecordState.kt
상세보기 MVI 패턴 viewModel과 state, event, effect 추가.

4) WishBookEntity.kt, WishBookModel.kt, MyBookEntity.kt, MyBookModel.kt
itemId String -> Long 타입으로 변경.

5) Android Os의 시스템 정보 유지하기
Theme.kt, Scaffold.kt, MainActivity.kt
영역 및 표시 설정 추가.

+++ data/book/src/main/java/lee/project/data/book/remote/dto/BookMapper.kt
@@ -27,7 +27,7 @@ fun AladinResponse<BookDto>.toDomain(): BookListModel {
 
 fun BookDto.toDomain(): BookModel {
     return BookModel(
-        itemId = this.itemId,
+        itemId = this.itemId.toLong(),
         title = this.title,
         link = this.link,
         author = this.author,
